TYPE REFACTOR - 1
