+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/evaluationResults.ts
@@ -4,165 +4,13 @@ import {WB_RUN_COLORS} from '../../../../../../common/css/color.styles';
 import {parseRef, WeaveObjectRef} from '../../../../../../react';
 import {PREDICT_AND_SCORE_OP_NAME_POST_PYDANTIC} from '../common/heuristics';
 import {
-  TraceCallSchema,
-  TraceObjSchema,
   TraceServerClient,
 } from '../wfReactInterface/traceServerClient';
 import {
   convertISOToDate,
   projectIdFromParts,
 } from '../wfReactInterface/tsDataModelHooks';
-import {EvaluationEvaluateCallSchema} from './evaluations';
-import {EvaluationComparisonState} from './types';
-
-type BinarySummaryScore = {
-  true_count: number;
-  true_fraction: number;
-};
-
-type ContinuousSummaryScore = {
-  mean: number;
-};
-
-export const isBinarySummaryScore = (
-  score: any
-): score is BinarySummaryScore => {
-  return (
-    typeof score === 'object' &&
-    score != null &&
-    'true_count' in score &&
-    'true_fraction' in score
-  );
-};
-
-export const isContinuousSummaryScore = (
-  score: any
-): score is ContinuousSummaryScore => {
-  return typeof score === 'object' && score != null && 'mean' in score;
-};
-
-type EvaluationCall = {
-  callId: string;
-  name: string;
-  color: string;
-  evaluationRef: string;
-  modelRef: string;
-  scores: {
-    [scoreName: string]: {
-      [path: string]: BinarySummaryScore | ContinuousSummaryScore;
-    };
-  };
-  _rawEvaluationTraceData: EvaluationEvaluateCallSchema;
-};
-
-type EvaluationObj = {
-  ref: string;
-  datasetRef: string;
-  scorerRefs: string[];
-  project: string;
-  entity: string;
-  _rawEvaluationObject: TraceObjSchema;
-};
-
-//   type Scorer = {
-//     ref: string;
-//     //   scoreSchema: string; ??
-//     _rawScorerObjectVersionData: TraceObjSchema;
-//   };
-
-//   type Dataset = {
-//     ref: string;
-//     rowDigests: string[];
-//     _rawDatasetObjectVersionData: TraceObjSchema;
-//   };
-
-// type PredictCall = {
-//   callId: string;
-//   inputDigest: string;
-//   modelRef: string;
-//   _rawPredictTraceData: TraceCallSchema;
-// };
-
-type DatasetRow = {
-  digest: string;
-  val: any;
-};
-
-type ModelObj = {
-  ref: string;
-  predictOpRef: string;
-  properties: {[prop: string]: any};
-  project: string;
-  entity: string;
-  _rawModelObject: TraceObjSchema;
-};
-
-type ScoreResults = {
-  callId: string;
-  results: {[path: string]: number | boolean};
-  _rawScoreTraceData: TraceCallSchema;
-};
-
-export type EvaluationComparisonData = {
-  entity: string;
-  project: string;
-  evaluationCalls: {
-    [callId: string]: EvaluationCall;
-  };
-  evaluations: {
-    [objectRef: string]: EvaluationObj;
-  };
-  //   datasets: {
-  //     [objectRef: string]: Dataset;
-  //   };
-  //   scorers: {
-  //     [objectRef: string]: Scorer;
-  //   };
-  inputs: {
-    [rowDigest: string]: DatasetRow;
-  };
-  models: {
-    [modelRef: string]: ModelObj;
-  };
-  resultRows: {
-    [rowDigest: string]: {
-      evaluations: {
-        [evaluationCallId: string]: {
-          predictAndScores: {
-            [predictAndScoreCallId: string]: PredictAndScoreCall;
-          };
-        };
-      };
-    };
-  };
-};
-
-type PredictAndScoreCall = {
-  callId: string;
-  firstExampleRef: string;
-  rowDigest: string;
-  modelRef: string;
-  evaluationCallId: string;
-  predictCall?: {
-    callId: string;
-    output: any;
-    latencyMs: number;
-    totalUsageTokens: number;
-    _rawPredictTraceData: TraceCallSchema;
-  };
-  scores: {
-    [scorerRef: string]: ScoreResults;
-  };
-  _rawPredictAndScoreTraceData: TraceCallSchema;
-};
-
-// const generateColorFromId = (id: string) => {
-//   const hash = id.split('').reduce((acc, char) => acc + char.charCodeAt(0), 0);
-//   const hue = hash % 360;
-//   const saturation = 70 + (hash % 30); // Saturation between 70% and 100%
-//   const lightness = 40 + (hash % 20); // Lightness between 40% and 60%
-//   return `hsl(${hue}, ${saturation}%, ${lightness}%)`;
-// };
+import {EvaluationComparisonData, EvaluationComparisonState, EvaluationEvaluateCallSchema, isBinarySummaryScore, isContinuousSummaryScore } from './ecpTypes';
 
 const pickColor = (ndx: number) => {
   return WB_RUN_COLORS[ndx % WB_RUN_COLORS.length];
