--- weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/types.ts
@@ -1,11 +1,0 @@
-import {EvaluationComparisonData} from './evaluationResults';
-import {ScoreDimension} from './evaluations';
-import {RangeSelection} from './initialize';
-
-export type EvaluationComparisonState = {
-  data: EvaluationComparisonData;
-  baselineEvaluationCallId: string;
-  comparisonDimension: ScoreDimension;
-  rangeSelection: RangeSelection;
-  selectedInputDigest?: string;
-};
